18_docker/fastapi_app/requirements.txt
fastapi
uvicorn
pinecone-client
